// test/client/utils/ErrorMsgs.ts
export const enum ErrorMsgs {
    EVENT_REQUIRED = 'Event is required',
    INVALID_IDENTITY = 'Invalid Identity provided.  A valid Identity contains both a uuid and name',
    INVALID_TARGET_WINDOW = 'Invalid targetWindow provided',
    INVALID_WINDOW = 'Invalid window provided',
    PROPERTIES_REQUIRED = 'Properties are required'
}